Turn components list into sceneobjectcollection

+++ JRPG_BattleSystem/Game/include/MainMenu/MainMenuScene.h
@@ -10,7 +10,7 @@ public:
 	void CreateScene() override;
 
 private:
-	void CreateLoadSceneButton(glm::vec2 position, glm::vec2 buttonSize, std::string sceneToLoad, std::string content, int textSize);
+	void CreateLoadSceneButton(glm::vec2 position, glm::vec2 buttonSize, std::string sceneToLoad, std::string content, float textSize);
 };
 
 #endif // __MAIN_MENU_H_INCLUDED__
